+++ CHANGES.md
@@ -1,5 +1,15 @@
+# 1.2.4
+January 2021
+
+- Namespace for cues for tools changed from 
+`http://www.clarin.eu/cmdi/cues/1` to `http://www.clarin.eu/cmd/cues/1`
+(see [#14](https://github.com/clarin-eric/cmdi-toolkit/issues/14))
+- Added Schematron rules for best practices
+- Group ID for Maven artifact changed to `eu.clarin.cmd`
+
 # 1.2.3
 November 2019
+
 - Updated Java libraries:
   - SchemAnon 1.0.0 -> 1.1.0
   - cmdi-validator-core 1.0.0 -> 1.2.1
@@ -7,6 +17,7 @@ November 2019
 
 # 1.2.2
 June 2017
+
 - Fixed some issues related to the creation of vocabulary information in the profile XSDs
 
 # 1.2.1